[refactor] calculator as a service
