+++ src/configs/t-line-calculator.config/t-line-calculator.config.service.ts
@@ -7,7 +7,9 @@ export class TLineCalculatorConfigService {
        S = SCORE
        W = WEIGHT
        C = CONST
+       F = FUNCTION
      */
+
     SW_SEC_REL = 2;
     SW_AUTHOR_PER = 1.5;
     SW_AUTHOR_REL = 1.5;
@@ -17,5 +19,8 @@ export class TLineCalculatorConfigService {
 
     C_ATAN_DIVISOR = 50;
 
+    C_IDEAL_POSTS_PER_SEC = 3;
 
+    // desmos code y=\frac{1}{x+4}\ +\ 0.75
+    F_SEEN_WEIGHT = (x) => (1 / (x + 4)) + 0.75
 }
