+++ src/modules/app.module.ts
@@ -1,8 +1,9 @@
 import {Module} from '@nestjs/common';
 import {TLineModule} from './t-line/t-line.module'
+import { TLineCalculatorModule } from './t-line-calculator/t-line-calculator.module';
 
 @Module({
-    imports: [TLineModule],
+    imports: [TLineModule, TLineCalculatorModule],
 })
 export class AppModule {
 }
