+++ src/modules/t-line/t-line.controller.spec.ts
@@ -1,5 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { TLineController } from './t-line.controller';
+import {describe, expect, it} from "@jest/globals";
 
 describe('TLineController', () => {
   let controller: TLineController;
